Fix : refactor CategorySideBar
