+++ src/components/Home/index.tsx
@@ -1,28 +1,10 @@
-import { useState } from 'react';
-import CategorySideBar from './CategorySideBar';
 import ItemList from './ItemList';
 import styles from './index.module.css';
-import togglebutton from '../../resources/image/menu.svg';
 
 export default function HomePage() {
-  const [openCategorySideBar, setopenCategorySideBar] = useState(false);
-
-  const toggleCategorySideBar = () => {
-    setopenCategorySideBar(!openCategorySideBar);
-  };
-
   return (
-    <div className={styles.container}>
-      {openCategorySideBar && <CategorySideBar />}
-      <div className={styles.mainContent}>
-        <img
-          className={styles.openCategorySideBarButton}
-          onClick={toggleCategorySideBar}
-          src={togglebutton}
-          alt="카테고리더보기"
-        ></img>
-        <ItemList />
-      </div>
+    <div className={styles.mainContent}>
+      <ItemList />
     </div>
   );
 }
